+++ gbif/src/main/java/org/gbif/pipelines/interpretation/column/CountryCodeInterpreter.java
@@ -1,4 +1,4 @@
-package org.gbif.pipelines.core.functions.interpretation;
+package org.gbif.pipelines.interpretation.column;
 
 import org.gbif.api.vocabulary.Country;
 import org.gbif.common.parsers.CountryParser;
@@ -12,26 +12,24 @@ import java.util.Collections;
 import java.util.List;
 
 /**
- * interprets country
+ * interprets country code
  */
-class CountryInterpreter implements Interpretable<String> {
+class CountryCodeInterpreter implements Interpretable<String,String> {
 
     @Override
-    public InterpretationResult<String> interpret(String input) {
+    public InterpretationResult<String> apply(String input) {
         ParseResult<Country> parseCountry = CountryParser.getInstance().parse(input.trim());
         if (parseCountry.isSuccessful()) {
-            return InterpretationResult.withSuccess(parseCountry.getPayload().getTitle());
+            return InterpretationResult.withSuccess(parseCountry.getPayload().getIso3LetterCode());
         } else {
             final List<Issue> issues = Collections.singletonList(Issue.newBuilder()
                     .setIssueType(IssueType.PARSE_ERROR)
-                    .setRemark(parseCountry.getError() != null
-                            ? parseCountry.getError().getMessage()
-                            : "Parse failed as null")
+                    .setRemark(parseCountry.getError() != null ? parseCountry.getError().getMessage() : "Couldnot parse countryCode")
                     .build());
             final List<Lineage> lineages = Collections.singletonList(Lineage.newBuilder()
                     .setLineageType(LineageType.SET_TO_NULL)
                     .setRemark(
-                            "Since the parse on country failed, interpreting as null")
+                            "Since the parse on countryCode failed, interpreting as null")
                     .build());
             return InterpretationResult.withIssueAndLineage(null, issues, lineages);
         }
