+++ gbif/src/test/java/org/gbif/pipelines/transforms/InterpretedCategoryTransformerTest.java
@@ -18,8 +18,8 @@ import org.gbif.dwca.avro.Event;
 import org.gbif.dwca.avro.ExtendedOccurence;
 import org.gbif.dwca.avro.Location;
 import org.gbif.pipelines.common.beam.Coders;
-import org.gbif.pipelines.core.functions.interpretation.InterpretationFactory;
-import org.gbif.pipelines.core.functions.interpretation.InterpretationResult;
+import org.gbif.pipelines.interpretation.column.InterpretationFactory;
+import org.gbif.pipelines.interpretation.column.InterpretationResult;
 import org.gbif.pipelines.core.functions.interpretation.error.Issue;
 import org.gbif.pipelines.core.functions.interpretation.error.IssueLineageRecord;
 import org.gbif.pipelines.core.functions.interpretation.error.Lineage;
